refactoring to address codacy points, tidy up naming
