+++ handlers/identity-backfill/src/main/scala/com/gu/identityBackfill/zuora/GetZuoraAccountsForEmail.scala
@@ -2,7 +2,8 @@ package com.gu.identityBackfill.zuora
 
 import com.gu.identityBackfill.Types._
 import com.gu.util.zuora.RestRequestMaker.ClientFailableOp
-import com.gu.util.zuora.ZuoraQuery.{Query, ZuoraQuerier}
+import com.gu.util.zuora.SafeQueryBuilder.Implicits._
+import com.gu.util.zuora.ZuoraQuery.ZuoraQuerier
 import play.api.libs.json.Json
 import scalaz.ListT
